+++ CHANGELOG.md
@@ -16,3 +16,36 @@
 - Change Old Structure
 
 ### Fixed
+
+## [4.2.0] - 2023-01-17
+
+### Added
+
+- Add New methods to Product CRUD functionality
+
+### Changed
+
+- Change Structure classes to prepare data
+
+### Fixed
+
+## [4.3.0] - 2023-01-18
+
+### Added
+
+- Code Review
+- Add New Methods to CRUD functionality
+
+## [4.3.1] - 2023-01-18
+
+### Changed
+
+- Remove Legacy Code
+
+## [4.3.2] - 2023-01-30
+
+### Changed
+
+- Set attributes prepare and setup methods
+- Set Image data
+- Set default variations methods